Bugfixes to prevent endless loops and added realtime logging. (#1)
* Fixed a bug that prevents the tool from running in an endless loop at 100% CPU usage when faulty input data is provided. When a fault is now detected, the tool will stopp with an error message.
* The tool now also ends up in an error if the expected input data format is not found.
* Modified the Python code to include realtime logging. Before this change, logging was only printed after a process finished. If the process never finished, it meant no log data to know what went wrong.

+++ make_otu_table.sh
@@ -4,16 +4,16 @@ outlocation=$(mktemp -d /data/files/XXXXXX)
 SCRIPTDIR=$(dirname "$(readlink -f "$0")")
 
 # sanity check
-printf "Conda env: $CONDA_DEFAULT_ENV\n"
-printf "Outlocation: $outlocation\n"
-printf "Python version: $(python --version |  awk '{print $2}')\n"
-printf "Biopython version: $(conda list | egrep biopython | awk '{print $2}')\n"
-printf "Usearch version: $(usearch11 | head -n1 | awk '{print $2}' | tr -d ',')\n"
-printf "Vsearch version: $(conda list | egrep vsearch | awk '{print $2}')\n"
-printf "Dada2 version: $(conda list | egrep dada2 | awk '{print $2}')\n"
-printf "Unzip version: $(unzip -v | head -n1 | awk '{print $2}')\n"
-printf "Bash version: ${BASH_VERSION}\n"
-printf "SCRIPTDIR: $SCRIPTDIR\n\n"
+#printf "Conda env: $CONDA_DEFAULT_ENV\n"
+#printf "Outlocation: $outlocation\n"
+#printf "Python version: $(python --version |  awk '{print $2}')\n"
+#printf "Biopython version: $(conda list | egrep biopython | awk '{print $2}')\n"
+#printf "Usearch version: $(usearch11 | head -n1 | awk '{print $2}' | tr -d ',')\n"
+#printf "Vsearch version: $(conda list | egrep vsearch | awk '{print $2}')\n"
+#printf "Dada2 version: $(conda list | egrep dada2 | awk '{print $2}')\n"
+#printf "Unzip version: $(unzip -v | head -n1 | awk '{print $2}')\n"
+#printf "Bash version: ${BASH_VERSION}\n"
+#printf "SCRIPTDIR: $SCRIPTDIR\n\n"
 
 if [ $3 == "cluster_otus" ]
 then
@@ -44,11 +44,6 @@ then
 python $SCRIPTDIR"/make_otu_table.py" -i $1 -t $2 -c $3 -of $outlocation -cluster_id ${9} -abundance_minsize "${10}" -cluster_size "${11}"
 fi
 
-#usearch11 -otutab_stats $outlocation"/otutab.txt" -output $outlocation/"report.txt" &> /dev/null
-#echo "Otu table summary" >> $outlocation"/log.log"
-#echo "============================================================" >> $outlocation"/log.log"
-#cat $outlocation/"report.txt" >> $outlocation"/log.log"
-
 #output files
 if [ $4 ]
 then
@@ -64,7 +59,7 @@ if [ $6 ]
 then
 	cp $outlocation"/otu_sequences.fa" $6 && [ -f $outlocation"/otu_sequences.fa" ]
 	# convert interleaved or multiline fasta to singleline
-	cat $outlocation"/otu_sequences.fa" | 
+	cat $outlocation"/otu_sequences.fa" |
 	awk '/^>/ { if(NR>1) print "";  printf("%s\n",$0); next; } { printf("%s",$0);}  END {printf("\n");}' > $outlocation"/otu_sequences_DG.fa"
 	rm $outlocation"/otu_sequences.fa"
 	#------------------------------------------------------------------
@@ -72,13 +67,33 @@ then
 	#  Files seem to be sorted, but sort anyway, just in case.
 	#------------------------------------------------------------------
 	# max length Otu label
-	max_length=$(cat $outlocation"/otu_sequences_DG.fa"| egrep "^>Otu" | awk '{print length($1)}' | sort -n | uniq | tail -n1) 
+
+  # Check if file exists. If not, then something went wrong and the input data might be incorrect.
+  if [ ! -e $outlocation"/otu_sequences_DG.fa" ]; then
+      echo "ERROR: /otu_sequences_DG.fa file does not exist. Please contact your Galaxy admin" >&2
+  fi
+
+  # Check if file is empty. If so, then something went wrong and the input data might be incorrect.
+  if [ ! -s $outlocation"/otu_sequences_DG.fa" ]; then
+      echo "ERROR: /otu_sequences_DG.fa file is empty. Please contact your Galaxy admin" >&2
+      exit 1
+  fi
+
+	max_length=$(cat $outlocation"/otu_sequences_DG.fa"| egrep "^>Otu" | awk '{print length($1)}' | sort -n | uniq | tail -n1)
 	# max number of digits of Otu label (substract "Otu" from max_length Otu label)
 	otu_digits=$(echo $max_length-4 | bc)
-	# create a string of zeros 
+
+	# Perform an additional check to make sure that $otu_digits is not a negative number. This would make the script
+  # end up in an infinite loop.
+	if (( $otu_digits < 1 )); then
+      echo "ERROR: otu_digits is less than 1! Please contact your Galaxy admin" >&2
+      exit 1
+	fi
+
+	# create a string of zeros
 	otu_digit_string=$(echo $(yes "0" | head -n "$otu_digits") | tr -d " ")
 	# padding zeros
-	cat $outlocation"/otu_sequences_DG.fa" | 
+	cat $outlocation"/otu_sequences_DG.fa" |
 	sed "s/\(^>Otu\)\([0-9]\)/\1$otu_digit_string\2/g; s/0*\([0-9]\{$otu_digits,\}\)/\1/g" | paste - - | sort -n | sed 's/\t/\n/g' > 	$outlocation"/otu_sequences_DG2.fa"
 	rm $outlocation"/otu_sequences_DG.fa"
 	cp $outlocation"/otu_sequences_DG2.fa" $6 && [ -f $outlocation"/otu_sequences_DG2.fa" ]
@@ -94,10 +109,29 @@ then
 	#  adjust Otu label format and sort
 	#------------------------------------------------------------------
 	# max length Otu label
-	max_length=$(cat $outlocation"/otutab.txt" | egrep "^Otu" | awk '{print length($1)}' | sort -n | uniq | tail -n1) 
-	# max number of digits of Otu label (substract "Otu" from max_length Otu label)
+
+	# Check if file exists. If not, then something went wrong and the input data might be incorrect.
+	if [ ! -e $outlocation"/otutab.txt" ]; then
+	    echo "ERROR: otutab.txt file does not exist. Please contact your Galaxy admin" >&2
+	fi
+
+  	# Check if file is empty. If so, then something went wrong and the input data might be incorrect.
+	if [ ! -s $outlocation"/otutab.txt" ]; then
+	    echo "ERROR: otutab.txt file is empty. Please contact your Galaxy admin" >&2
+	    exit 1
+	fi
+
+	max_length=$(cat $outlocation"/otutab.txt" | egrep "^Otu" | awk '{print length($1)}' | sort -n | uniq | tail -n1)
+	## max number of digits of Otu label (substract "Otu" from max_length Otu label)
 	otu_digits=$(echo $max_length-3 | bc)
-	# create a string of zeros 
+  # Perform an additional check to make sure that $otu_digits is not a negative number. This would make the script
+  # end up in an infinite loop.
+	if (( $otu_digits < 1 )); then
+      echo "ERROR: otu_digits is less than 1! Please contact your Galaxy admin" >&2
+      exit 1
+  fi
+
+	# create a string of zeros
 	otu_digit_string=$(echo $(yes "0" | head -n "$otu_digits") | tr -d " ")
 	# padding zeros
 	cat $outlocation"/otutab.txt" | sed "s/\(^Otu\)\([0-9]\)/\1$otu_digit_string\2/g; s/0*\([0-9]\{$otu_digits,\}\)/\1/g" | sort -n > $outlocation"/otutab_DG.txt"
@@ -111,3 +145,4 @@ then
     mv $outlocation"/bioom.json" $8 && [ -f $outlocation"/bioom.json" ]
 fi
 rm -rf $outlocation
+echo "script finished"
